requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Optimization libraries
cvxpy>=1.2.0
pymoo>=0.6.0

# Machine Learning
scikit-learn>=1.0.0
torch>=1.12.0
tensorflow>=2.8.0

# Transformers and NLP
transformers>=4.20.0
huggingface-hub>=0.8.0
tokenizers>=0.12.0

# Portfolio optimization specific
PyPortfolioOpt>=1.5.0
riskfolio-lib>=3.0.0
deepdow>=1.0.0

# Quantum computing
qutip>=4.7.0
qiskit>=0.45.0
qiskit-aer>=0.12.0
qiskit-optimization>=0.6.0
qiskit-machine-learning>=0.7.0
cirq>=1.2.0
pennylane>=0.32.0
pennylane-qiskit>=0.32.0
dwave-ocean-sdk>=6.0.0
dwave-system>=1.19.0
dwave-networkx>=0.8.0

# Real-time data and streaming
kafka-python>=2.0.0
redis>=4.5.0
websocket-client>=1.6.0
alpha-vantage>=2.3.0
polygon-api-client>=1.12.0

# High-performance computing
cupy-cuda11x>=12.0.0  # For NVIDIA GPUs with CUDA 11.x
numba>=0.58.0
ray>=2.7.0
dask>=2023.9.0
rapids-cudf>=23.10.0  # Optional: RAPIDS for GPU DataFrames

# Advanced ML and AI
torch-geometric>=2.4.0
transformers[torch]>=4.34.0
sentence-transformers>=2.2.0
datasets>=2.14.0
accelerate>=0.23.0

# Financial data and APIs - REAL-TIME ENHANCED
yfinance>=0.2.0
yflive>=0.1.0  # Real-time Yahoo Finance data
finnhub-python>=2.4.0  # Real-time market data
alpha-vantage>=2.3.0
quandl>=3.7.0
fredapi>=0.5.0
iexfinance>=0.5.0
polygon-api-client>=1.12.0
tiingo>=0.14.0

# News and Sentiment Analysis - REAL-TIME
newsapi-python>=0.2.6  # Real-time news
newspaper3k>=0.2.8  # News scraping
textblob>=0.17.1  # Sentiment analysis
vaderSentiment>=3.3.2  # Social media sentiment
tweepy>=4.14.0  # Twitter API
praw>=7.7.0  # Reddit API
requests-html>=0.10.0  # Web scraping
beautifulsoup4>=4.12.0  # HTML parsing
feedparser>=6.0.10  # RSS feeds

# Real-time WebSocket and Streaming
websocket-client>=1.6.0
python-socketio>=5.9.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0
paho-mqtt>=1.6.0

# Web framework and APIs
fastapi>=0.103.0
uvicorn>=0.23.0
streamlit>=1.27.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
websockets>=11.0.0
plotly>=5.17.0

# Database and storage
sqlalchemy>=2.0.0
redis>=4.6.0
pymongo>=4.5.0
influxdb-client>=1.38.0

# Monitoring and observability
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Reinforcement Learning
gym>=0.21.0
stable-baselines3>=1.6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Data handling and file processing
openpyxl>=3.0.0
xlrd>=2.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Configuration and CLI
pyyaml>=6.0
click>=8.0.0
tqdm>=4.64.0

# Revolutionary AI and Optimization
optuna>=3.4.0  # Neural Architecture Search
ray[rllib]>=2.7.0  # Distributed RL and hyperparameter optimization
stable-baselines3>=2.0.0  # Advanced RL algorithms
gym>=0.26.0  # RL environments

# Blockchain and DeFi
web3>=6.11.0  # Ethereum integration
ccxt>=4.1.0  # Cryptocurrency exchange integration
requests>=2.31.0  # HTTP requests for DeFi APIs

# Edge Computing and Ultra-Low Latency
zmq>=0.0.0  # ZeroMQ for ultra-low latency messaging
pyzmq>=25.1.0  # Python ZeroMQ bindings
mmap>=0.0.0  # Memory mapping for zero-copy operations

# Advanced Numerical Computing
numba>=0.58.0  # JIT compilation for ultra-fast optimization
llvmlite>=0.41.0  # LLVM for Numba

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
jupyter>=1.0.0  # For analysis notebooks